Move distributed counter samples and tests into separate files.
